requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
pymongo
pandas
scikit-learn
joblib
requests
qrcode
streamlit
